test(proxy): split proxy session contract targets
