+++ Tests/ProxyStartupLoggingTests/ToolCatalogStartupLogFormatterTests.swift
